(fix) evaluate rule failing & dataset filters rendering
