+++ frontend/src/sections/annotations/queues/constants.js
@@ -1,3 +1,5 @@
+import { SIMULATION_PERSONA_FILTER_FIELDS } from "./utils/simulation-persona-filter-fields";
+
 export const QUEUE_ROLES = {
   ANNOTATOR: "annotator",
   REVIEWER: "reviewer",
@@ -42,3 +44,112 @@ export const queueViewerMembership = (queue) => {
 
 export const canViewerAddItemsToQueue = (queue) =>
   hasQueueRole(queueViewerMembership(queue), QUEUE_ROLES.MANAGER);
+
+
+export const SOURCE_OPTIONS = [
+  { value: "dataset_row", label: "Dataset Row" },
+  { value: "trace", label: "Trace" },
+  { value: "observation_span", label: "Span" },
+  { value: "trace_session", label: "Session" },
+  { value: "call_execution", label: "Simulation" },
+];
+
+export const TRIGGER_FREQUENCY_OPTIONS = [
+  { value: "manual", label: "Manually" },
+  { value: "hourly", label: "Every hour" },
+  { value: "daily", label: "Daily" },
+  { value: "weekly", label: "Weekly" },
+  { value: "monthly", label: "Monthly" },
+];
+
+export const DEFAULT_FILTER = {
+  column_id: "",
+  filter_config: {
+    filter_type: "",
+    filter_op: "",
+    filter_value: "",
+  },
+};
+
+export const SIMULATION_RULE_FILTER_FIELDS = [
+  {
+    id: "status",
+    name: "Status",
+    category: "system",
+    type: "categorical",
+    choices: ["completed", "failed", "in_progress", "pending", "cancelled"],
+  },
+  ...SIMULATION_PERSONA_FILTER_FIELDS,
+  {
+    id: "agent_definition",
+    name: "Agent Definition",
+    category: "system",
+    type: "text",
+  },
+  {
+    id: "call_type",
+    name: "Call Type",
+    category: "system",
+    type: "categorical",
+    choices: ["voice", "text"],
+  },
+  {
+    id: "simulation_call_type",
+    name: "Simulation Call Type",
+    category: "system",
+    type: "text",
+  },
+  {
+    id: "duration_seconds",
+    name: "Duration",
+    category: "system",
+    type: "number",
+  },
+  {
+    id: "overall_score",
+    name: "Overall Score",
+    category: "system",
+    type: "number",
+  },
+  {
+    id: "created_at",
+    name: "Created At",
+    category: "system",
+    type: "date",
+  },
+];
+
+export const SIMPLE_FILTER_CATEGORIES = [
+  { key: "all", label: "All", icon: "mdi:view-grid-outline" },
+  { key: "system", label: "System", icon: "mdi:tune-variant" },
+  { key: "persona", label: "Persona", icon: "mdi:account-outline" },
+];
+
+export const SESSION_RULE_FILTER_FIELDS = [
+  { id: "session_id", name: "Session ID", category: "system", type: "string" },
+  {
+    id: "first_message",
+    name: "First Message",
+    category: "system",
+    type: "string",
+  },
+  {
+    id: "last_message",
+    name: "Last Message",
+    category: "system",
+    type: "string",
+  },
+  { id: "user_id", name: "User ID", category: "system", type: "string" },
+  { id: "duration", name: "Duration", category: "system", type: "number" },
+  { id: "total_cost", name: "Total Cost", category: "system", type: "number" },
+  {
+    id: "total_traces_count",
+    name: "Total Traces",
+    category: "system",
+    type: "number",
+  },
+  { id: "start_time", name: "Start Time", category: "system", type: "date" },
+  { id: "end_time", name: "End Time", category: "system", type: "date" },
+];
+
+export const MULTI_VALUE_OPS = new Set(["in", "not_in"]);
