+++ frontend/src/sections/projects/LLMTracing/TraceFilterPanel.jsx
@@ -1550,7 +1550,7 @@ function FilterRow({
       );
     }
 
-    if (filter.fieldType === "text") {
+    if (filter.fieldType === "text" || filter.fieldType === "string") {
       return (
         <TextField
           size="small"
